+++ morsecode/functions.py
@@ -12,18 +12,17 @@ It defines:
 8. and clear_all().
 """
 
+import re
+import os
 import tkinter as tk
-from tkinter import messagebox as msgbox
-from os.path import splitext, exists
-from os import makedirs
-from playsound3 import playsound
-from playsound3.playsound3 import PlaysoundException
+from tkinter import messagebox
+import playsound3
 import unidecode
 import pyperclip
 import pycw
-from modules.globals import *
+import globals
 
-# Path to the most recently generated audio file
+# Filepath to the most recently generated audio file
 most_recent_audio_filepath: str = ""
 
 
@@ -43,15 +42,15 @@ def play_audio_file(audio_status: str) -> None:
     # Check whether the audio_status is non-empty
     if audio_status:
         try:
-            playsound(sound=most_recent_audio_filepath)
+            playsound3.playsound(sound=most_recent_audio_filepath)
 
-        except PlaysoundException as e:
+        except playsound3.playsound3.PlaysoundException as e:
             # Display an error if the audio file can't be played
-            msgbox.showerror(title=MSGBOX_TITLE_ERROR, message=str(e))
+            messagebox.showerror(title=globals.MESSAGEBOX_TITLE_ERROR, message=str(e))
 
     else:
         # Display a warning if no audio file has been created yet
-        msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_PLAY_WARNING)
+        messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING, message=globals.MESSAGEBOX_MSG_PLAY_WARNING)
 
 
 def change_entry_text(entry_to_change: tk.Entry, change_to: str = "") -> None:
@@ -116,12 +115,13 @@ def create_audio_file(normalized_text: str, audio_status: tk.Entry) -> None:
 
     # Check whether the normalized_text contains any chars unsupported
     # by the pycw module.
-    if bool(re.search(pattern=PATT_SANITIZE_PLAIN_TEXT_INPUT, string=normalized_text)):
+    if bool(re.search(pattern=globals.PATT_SANITIZE_PLAIN_TEXT_INPUT, string=normalized_text)):
 
         # If so, ask the user whether to automatically remove them
         # from the text, or whether to abandon audio generation.
-        if msgbox.askyesno(title=MSGBOX_TITLE_CONFIRM, message=MSGBOX_MSG_TRANSLATE_TO_MORSE_CONFIRM):
-            normalized_text = re.sub(pattern=PATT_SANITIZE_PLAIN_TEXT_INPUT, repl="", string=normalized_text)
+        if messagebox.askyesno(title=globals.MESSAGEBOX_TITLE_CONFIRM,
+                               message=globals.MESSAGEBOX_MSG_TRANSLATE_TO_MORSE_CONFIRM):
+            normalized_text = re.sub(pattern=globals.PATT_SANITIZE_PLAIN_TEXT_INPUT, repl="", string=normalized_text)
 
         else:
             change_most_recent_filepath()
@@ -131,13 +131,13 @@ def create_audio_file(normalized_text: str, audio_status: tk.Entry) -> None:
     # Incrementally, find the first unused audio filename to avoid
     # overwriting previous ones.
     counter: int = 0
-    audio_filepath: str = f"{DEFAULT_AUDIO_OUTPUT_DIR}/{DEFAULT_AUDIO_OUTPUT_FILE}"
-    audio_file_name, audio_file_ext = splitext(p=audio_filepath)
+    audio_filepath: str = f"{globals.DEFAULT_AUDIO_OUTPUT_DIR}/{globals.DEFAULT_AUDIO_OUTPUT_FILE}"
+    audio_file_name, audio_file_ext = os.path.splitext(p=audio_filepath)
 
     # Create the directory if it doesn't exist yet
-    makedirs(name=DEFAULT_AUDIO_OUTPUT_DIR, exist_ok=True)
+    os.makedirs(name=globals.DEFAULT_AUDIO_OUTPUT_DIR, exist_ok=True)
 
-    while exists(audio_filepath):
+    while os.path.exists(audio_filepath):
         counter += 1
         audio_filepath = f"{audio_file_name}_{counter}{audio_file_ext}"
 
@@ -145,20 +145,22 @@ def create_audio_file(normalized_text: str, audio_status: tk.Entry) -> None:
     try:
         pycw.output_wave(file=audio_filepath,
                          text=normalized_text,
-                         tone=AUDIO_TONE,
-                         volume=AUDIO_VOLUME,
-                         sample_rate=AUDIO_SAMPLE_RATE,
-                         wpm=AUDIO_WPM)
+                         tone=globals.AUDIO_TONE,
+                         volume=globals.AUDIO_VOLUME,
+                         sample_rate=globals.AUDIO_SAMPLE_RATE,
+                         wpm=globals.AUDIO_WPM)
 
     except PermissionError:
         # Display an error if the audio file can't be created
         # due to lack of permission.
-        msgbox.showerror(title=MSGBOX_TITLE_ERROR, message=MSGBOX_MSG_TRANSLATE_TO_MORSE_ERROR)
+        messagebox.showerror(title=globals.MESSAGEBOX_TITLE_ERROR,
+                             message=globals.MESSAGEBOX_MSG_TRANSLATE_TO_MORSE_ERROR)
 
     else:
         # Save the audio filepath as the most recent one
         change_most_recent_filepath(change_to=audio_filepath)
-        change_entry_text(entry_to_change=audio_status, change_to=f"{AUDIO_READY_MESSAGE} {most_recent_audio_filepath}")
+        change_entry_text(entry_to_change=audio_status,
+                          change_to=f"{globals.AUDIO_READY_MESSAGE} {most_recent_audio_filepath}")
 
 
 def translate_to_morse_code(user_plain_text: str, audio_request: bool, audio_status: tk.Entry,
@@ -187,7 +189,8 @@ def translate_to_morse_code(user_plain_text: str, audio_request: bool, audio_sta
         # Translate the normalized text to Morse code.
         # Each letter is looked up in the MORSE_CODE_DICT,
         # and words are separated by a slash with spaces.
-        mcode: str = " / ".join(" ".join(map(lambda c: MORSE_CODE_DICT.get(c, ""), word)) for word in split_user_text)
+        mcode: str = " / ".join(" ".join(map(lambda c: globals.MORSE_CODE_DICT.get(c, ""), word)) for word
+                                in split_user_text)
         mcode = mcode.strip()
 
         if mcode:
@@ -205,11 +208,12 @@ def translate_to_morse_code(user_plain_text: str, audio_request: bool, audio_sta
             # about the meaningless sequence in the input.
             change_most_recent_filepath()
             change_entry_text(entry_to_change=audio_status)
-            msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_MEANINGLESS_TO_MORSE_WARNING)
+            messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING,
+                                   message=globals.MESSAGEBOX_MSG_MEANINGLESS_TO_MORSE_WARNING)
 
     else:
         # Display a warning if the user has not provided any input
-        msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_TRANSLATE_WARNING)
+        messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING, message=globals.MESSAGEBOX_MSG_TRANSLATE_WARNING)
 
 
 def translate_to_plain_text(user_morse_code_text: str, audio_request: bool,
@@ -229,22 +233,24 @@ def translate_to_plain_text(user_morse_code_text: str, audio_request: bool,
 
     if not user_morse_code_text:
         # Display a warning if the user has not provided any input
-        msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_TRANSLATE_WARNING)
+        messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING, message=globals.MESSAGEBOX_MSG_TRANSLATE_WARNING)
 
-    elif not bool(re.fullmatch(pattern=PATT_SANITIZE_MORSE_CODE_INPUT, string=user_morse_code_text)):
+    elif not bool(re.fullmatch(pattern=globals.PATT_SANITIZE_MORSE_CODE_INPUT, string=user_morse_code_text)):
         # Display a warning if the user has provided illegal chars
-        msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_TRANSLATE_TO_PLAIN_WARNING)
+        messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING,
+                               message=globals.MESSAGEBOX_MSG_TRANSLATE_TO_PLAIN_WARNING)
 
     else:
 
         # Replace each occurrence of two spaces or more with one space
-        user_morse_code_text_reduced: str = re.sub(pattern=PATT_REDUCE_SPACES_MORSE_CODE_INPUT, repl=" ",
+        user_morse_code_text_reduced: str = re.sub(pattern=globals.PATT_REDUCE_SPACES_MORSE_CODE_INPUT,
+                                                   repl=" ",
                                                    string=user_morse_code_text)
         # Split the user's input into a list of words
         split_user_morse_code_text: list = user_morse_code_text_reduced.split(sep=" / ")
 
         # Create the dictionary with switched keys and values
-        switched_morse_code_dict: dict = {v: k for k, v in MORSE_CODE_DICT.items()}
+        switched_morse_code_dict: dict = {v: k for k, v in globals.MORSE_CODE_DICT.items()}
 
         # Look up each Morse code letter in that dict, and join each
         # word with the separator being a space.
@@ -265,11 +271,12 @@ def translate_to_plain_text(user_morse_code_text: str, audio_request: bool,
             # about the meaningless sequence in the input.
             change_most_recent_filepath()
             change_entry_text(entry_to_change=audio_status)
-            msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_MEANINGLESS_TO_PLAIN_WARNING)
+            messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING,
+                                   message=globals.MESSAGEBOX_MSG_MEANINGLESS_TO_PLAIN_WARNING)
 
 
 def copy_to_clipboard(text_to_copy: str) -> None:
-    """A utility function that copies the arg to the clipboard.
+    """Copies the arg to the system clipboard.
 
     Args:
         text_to_copy: str: String from an entry widget.
@@ -282,11 +289,11 @@ def copy_to_clipboard(text_to_copy: str) -> None:
         # Copy text to the system clipboard and display
         # a success message to the user.
         pyperclip.copy(text=text_to_copy)
-        msgbox.showinfo(title=MSGBOX_TITLE_SUCCESS, message=MSGBOX_MSG_COPY_SUCCESS)
+        messagebox.showinfo(title=globals.MESSAGEBOX_TITLE_SUCCESS, message=globals.MESSAGEBOX_MSG_COPY_SUCCESS)
 
     else:
         # Display a warning if there is no input
-        msgbox.showwarning(title=MSGBOX_TITLE_WARNING, message=MSGBOX_MSG_COPY_WARNING)
+        messagebox.showwarning(title=globals.MESSAGEBOX_TITLE_WARNING, message=globals.MESSAGEBOX_MSG_COPY_WARNING)
 
 
 def clear_all(*args: tk.Entry) -> None: